requirements/base.txt
django==5.1.3
python-dotenv==1.0.1
djangorestframework==3.15.2
django-autoslug==1.9.9
django-countries==7.6.1
django-phonenumber-field==8.0.0
phonenumbers==8.13.50
drf-yasg==1.21.8
pillow==11.0.0
argon2-cffi==*
djoser==2.3.1
django-taggit==6.1.0
django-filter==24.3
django-celery-email==3.0.0
cloudinary==1.41.0
django-celery-beat==2.7.0
redis==5.2.0
celery==5.4.0
flower==2.0.1